requirements.txt
flask
selenium
webdriver-manager
requests
pymongo
python-dotenv
dnspython
Flask-Bcrypt